requirements.txt
# OCR
pytesseract>=0.3.10
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Image Processing
scikit-image>=0.21.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0 